refactor: 💡 Rename files and verbose types
