+++ packages/upset/src/components/AggregateRow.tsx
@@ -6,8 +6,8 @@ import { useRecoilValue } from 'recoil';
 import { visibleSetSelector } from '../atoms/config/visibleSetsAtoms';
 import { dimensionsSelector } from '../atoms/dimensionsAtom';
 import translate from '../utils/transform';
-import { CardinalityBar } from './CardinalityBars';
-import { DeviationBar } from './DeviationBars';
+import { CardinalityBar } from './CardinalityBar';
+import { DeviationBar } from './DeviationBar';
 import { Matrix } from './Matrix';
 
 /** @jsxImportSource @emotion/react */
