+++ packages/upset/src/components/CardinalityBar.tsx
